% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{sustain}
\alias{sustain}
\title{sustain}
\usage{
sustain(x, casefun, tolerance = 1)
}
\description{
rolls over vector, "sustaining" values that return TRUE through the case-function.
Values are sustained "tolerance" number of times.
}
