% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{sustest}
\alias{sustest}
\title{sustest}
\usage{
sustest(x, casefun, tolerance = 1, ignore.na = FALSE)
}
\description{
rolls over vector, staying at 1 until \code{tolerance} number
of zeroes has been encountered.
}
