% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{until}
\alias{until}
\title{crop}
\usage{
until(x, case = 1, replacement = 0, replaceMissing = TRUE)
}
\description{
simple function that alters a series by inserting replacement value 
before encountering the case value.
}
